refactor: Split oversized SSH config modules into smaller submodules
Split include.rs (1028 lines) and match_directive.rs (972 lines) into
modular submodules for better maintainability and readability.

include module (609 + 266 + 192 lines):
- include/mod.rs: Core Include directive processing
- include/resolver.rs: Pattern resolution and glob handling
- include/validation.rs: Security validation functions

match_directive module (509 + 490 lines):
- match_directive/mod.rs: Match condition logic and parsing
- match_directive/exec.rs: Exec command execution and validation

All tests pass (79 tests) and clippy warnings resolved.
